# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WeldSim Virtual Simulation Platform API
  description: API for the Ultrasonic Metal Welding Virtual Simulation Platform
  version: 2.0.0
servers:
  - url: http://localhost:8001/api/v2
    description: Local development
paths:
...
  /projects/{project_id}/geometries:
    parameters:
      - name: project_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: List geometry versions
      operationId: listGeometries
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/GeometryVersion' }
    post:
      summary: Create geometry version
      operationId: createGeometry
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GeometryCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeometryVersion' }
  /geometries/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get geometry version
      operationId: getGeometry
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeometryVersion' }
  /geometries/{id}/generate:
    post:
      summary: Run parametric geometry generation
      operationId: generateGeometry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeometryVersion' }
  /geometries/{id}/mesh:
    post:
      summary: Generate mesh for geometry
      operationId: generateMesh
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                element_size: { type: number, default: 2.0 }
                mesh_density: { type: string, default: medium }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeometryVersion' }
  /geometries/{id}/preview:
    get:
      summary: Get mesh preview data for VTK.js
      operationId: getGeometryPreview
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeshPreview' }
  /geometries/upload:
    post:
      summary: Upload STEP/STL file
      operationId: uploadGeometry
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                project_id:
                  type: string
                  format: uuid
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeometryVersion' }

  # --- Simulations ---
...
  /projects/{project_id}/comparisons:
    parameters:
      - name: project_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Create comparison
      operationId: createComparison
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ComparisonCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comparison' }
  /comparisons/{id}:
    get:
      summary: Get comparison with results
      operationId: getComparison
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comparison' }
  /comparisons/{id}/refresh:
    post:
      summary: Recompute comparison deltas
      operationId: refreshComparison
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comparison' }

  # --- Optimizations ---
  /simulations/{simulation_id}/optimize:
    post:
      summary: Start optimization study
      operationId: createOptimization
      parameters:
        - name: simulation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OptimizationCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OptimizationStudy' }
  /optimizations/{id}:
    get:
      summary: Get optimization status
      operationId: getOptimization
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OptimizationStudy' }
  /optimizations/{id}/iterations:
    get:
      summary: List iteration results
      operationId: listIterations
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IterationResult' }
  /optimizations/{id}/pareto:
    get:
      summary: Get Pareto front
      operationId: getPareto
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IterationResult' }
  /optimizations/{id}/pause:
    post:
      summary: Pause optimization
      operationId: pauseOptimization
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OptimizationStudy' }
  /optimizations/{id}/resume:
    post:
      summary: Resume optimization
      operationId: resumeOptimization
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OptimizationStudy' }

  # --- Materials ---
  /materials:
    get:
      summary: List materials
      operationId: listMaterials
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Material' }
    post:
      summary: Add custom material
      operationId: createMaterial
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaterialCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Material' }
  /materials/{id}:
    get:
      summary: Get material properties
      operationId: getMaterial
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Material' }
...
components:
  schemas:
...
    HornParams:
      type: object
      properties:
        horn_type: { type: string, enum: [flat, cylindrical, exponential, blade, stepped] }
        width_mm: { type: number }
        height_mm: { type: number }
        length_mm: { type: number }
        material: { type: string }
        knurl_type: { type: string }
        knurl_pitch_mm: { type: number }
        knurl_depth_mm: { type: number }
        chamfer_radius_mm: { type: number }
        edge_treatment: { type: string, enum: [none, chamfer, fillet, compound] }
    GeometryCreate:
      type: object
      properties:
        label: { type: string }
        source_type: { type: string, enum: [parametric, imported_step, imported_stl] }
        parametric_params: { $ref: '#/components/schemas/HornParams' }
        mesh_config: { type: object }
    GeometryVersion:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        version_number: { type: integer }
        label: { type: string }
        source_type: { type: string }
        parametric_params: { type: object }
        file_path: { type: string }
        mesh_file_path: { type: string }
        metadata_json: { type: object }
        created_at: { type: string, format: date-time }
    MeshPreview:
      type: object
      properties:
        vertices: { type: array, items: { type: array, items: { type: number } } }
        faces: { type: array, items: { type: array, items: { type: integer } } }
        scalar_field: { type: array, items: { type: number } }
        node_count: { type: integer }
        element_count: { type: integer }

    # --- Simulation ---
...
    ComparisonCreate:
      type: object
      required: [name, run_ids]
      properties:
        name: { type: string }
        description: { type: string }
        run_ids: { type: array, items: { type: string, format: uuid } }
        metric_names: { type: array, items: { type: string } }
        baseline_run_id: { type: string, format: uuid }
    ComparisonResultItem:
      type: object
      properties:
        run_id: { type: string, format: uuid }
        metric_name: { type: string }
        value: { type: number }
        delta_from_baseline: { type: number }
        delta_percent: { type: number }
    Comparison:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        run_ids: { type: array, items: { type: string, format: uuid } }
        metric_names: { type: array, items: { type: string } }
        baseline_run_id: { type: string, format: uuid }
        results: { type: array, items: { $ref: '#/components/schemas/ComparisonResultItem' } }
        created_at: { type: string, format: date-time }

    # --- Optimization ---
    DesignVariable:
      type: object
      required: [name]
      properties:
        name: { type: string }
        var_type: { type: string, enum: [continuous, categorical], default: continuous }
        min_value: { type: number }
        max_value: { type: number }
        step: { type: number }
        choices: { type: array, items: { type: string } }
    Constraint:
      type: object
      required: [metric, operator, value]
      properties:
        metric: { type: string }
        operator: { type: string, enum: ["<=", ">=", "==", within_percent] }
        value: { type: number }
        tolerance_pct: { type: number }
    Objective:
      type: object
      required: [metric, direction]
      properties:
        metric: { type: string }
        direction: { type: string, enum: [minimize, maximize] }
        weight: { type: number, default: 1.0 }
    OptimizationCreate:
      type: object
      required: [name, design_variables, objectives]
      properties:
        name: { type: string }
        strategy: { type: string, enum: [parametric_sweep, bayesian, genetic, manual_guided], default: parametric_sweep }
        design_variables: { type: array, items: { $ref: '#/components/schemas/DesignVariable' } }
        constraints: { type: array, items: { $ref: '#/components/schemas/Constraint' } }
        objectives: { type: array, items: { $ref: '#/components/schemas/Objective' } }
        max_iterations: { type: integer, default: 50 }
    OptimizationStudy:
      type: object
      properties:
        id: { type: string, format: uuid }
        simulation_case_id: { type: string, format: uuid }
        name: { type: string }
        strategy: { type: string }
        design_variables: { type: array, items: { type: object } }
        constraints: { type: array, items: { type: object } }
        objectives: { type: array, items: { type: object } }
        status: { type: string, enum: [pending, running, completed, paused] }
        total_iterations: { type: integer }
        completed_iterations: { type: integer }
        best_run_id: { type: string, format: uuid }
        pareto_front_run_ids: { type: array, items: { type: string, format: uuid } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    IterationResult:
      type: object
      properties:
        iteration: { type: integer }
        run_id: { type: string, format: uuid }
        parameters: { type: object }
        metrics: { type: object, additionalProperties: { type: number } }
        feasible: { type: boolean }

    # --- Material ---
    MaterialCreate:
      type: object
      required: [name, density_kg_m3, youngs_modulus_pa, poisson_ratio]
      properties:
        name: { type: string }
        category: { type: string }
        density_kg_m3: { type: number }
        youngs_modulus_pa: { type: number }
        poisson_ratio: { type: number }
        yield_strength_mpa: { type: number }
        thermal_conductivity: { type: number }
        specific_heat: { type: number }
        acoustic_impedance: { type: number }
        properties_json: { type: object }
    Material:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        category: { type: string }
        density_kg_m3: { type: number }
        youngs_modulus_pa: { type: number }
        poisson_ratio: { type: number }
        yield_strength_mpa: { type: number }
        thermal_conductivity: { type: number }
        specific_heat: { type: number }
        acoustic_impedance: { type: number }
        properties_json: { type: object }

    # --- Workflow ---
...